requirements.txt
datasets==2.19.2
matplotlib==3.8.4
moviepy==1.0.3
numpy==2.0.0
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
scikit_learn==1.4.2
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
wasabi==1.1.2
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
bitsandbytes-cuda110==0.26.0.post2
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.2
debugpy==1.8.1
decorator==4.4.2
dill==0.3.8
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mmsdk==1.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
triton==2.3.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
validators==0.29.0
wasabi==1.1.2
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
